AstraDB | Session Expiry | Chat History | Special Chat Agents (#16)
Multiple change to session management, chat agent creation and
invocation, and added AstraDB as vector store.

## Description:

- Added AstraDB as vector store for the knowledge base. Configuration
allows to switch between AstraDB and Upstash. Still left Upstash in the
app just in case we have issues with AstraDB.
- Correct response closing issue in `eventHandler` in `Server.ts`.
Noticed we were incorrectly closing the session when the **request** was
closed, instead of closing it when the **response** was done.
- Updated session management to implement session expiration after a
certain timeout (currently set to 15 minutes, configurable from
`AppConfig.api.session_expiry_time`
- So, if a no query is received from front-end for a certain session ID,
we delete that session from our `SessionManager`.
- Please note that the front-end can still make requests using the same
session ID. If the session for provided ID is not active, we just add a
new active session for provided ID.
- Enable to continue chat even after backend goes down. We were not able
to continue chat earlier in case the backend goes down between the chat
as the session ID sent by front-end was no longer present in backend.
Now, if front-end send a session ID, if its not present in the backend,
we don't throw an error, we instead use it to initialize a new session.
When chat agent makes an API call to retrieve chat messages, it sends
this session ID. If it is already present on the chat history database
(AstraDB or Upstash), the corresponding chat messages are returned.
Else, new record is created.
- Improved initial system prompt used to setup the chat agent, to ensure
it gives first preference to the chat history and then checks the
knowledge base. This ensures chat agent takes into account chat history
and not just generate new responses using knowledge base alone. Also,
modified system prompt to restrict queries to topics related to context,
i.e., chat agent won't be an open-ended chat agent like ChatGPT but
answer questions related to a specific context like University of
Ottawa.
- Removed the need to use `TestChatAgent.ts`, instead have setup the
ability to get a specialized chat agent. Right now since we don't have
any other chat agents, we only use a University of Ottawa chat agent.
- Chat agent to be used is hard coded in `Session.ts -
Session.setupChatAgent`. This can eventually be received either as a
param from the front-end, for example: `context: "uOttawa"`or can be
used on a specific endpoint like `/api/conversation/uottawa`
- Will allow us to support or add more chat agents based on context, for
example, a chat agent on University of Toronto data. Chat agents
configured under `src/ai/chat-agents/specialized-chat-agents/`
- Removing the use of `TestChatAgent` also fixes an issue with chat
history retrieval. It was creating a new session ID instead of using the
one provided by API.